requirements.txt
protobuf
grpcio
grpcio-tools
opencv-python
matplotlib
Pillow
